+++ tests/test_memcache.py
@@ -410,10 +410,10 @@ class test_ServerConnection(unittest.TestCase):
         self.assertEqual(
                 repr(server), '<ServerConnection to memcached://localhost/>')
 
-    def test_SelectorRestirOnDownServer(self):
+    def test_SelectorFractalHasher(self):
         memcache = memcached2.Memcache((
                 'memcached://localhost/', 'memcached://localhost/',),
-                selector=memcached2.SelectorRestirOnDownServer())
+                selector=memcached2.SelectorFractalHasher())
 
         for i in range(10):
             memcache.set(str(i), '*' * i)